+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [4.3.0](https://github.com/riodelphino/macime/compare/v4.2.0...v4.3.0) (2026-02-21)
 
+* **feat:** Add `daemon set` and `daemon set log-level`
 * **feat:** Add `--cjk-delay` option
 * **docs:** Refine doc comments
 * **refactor:** Split colorizing to func in `Log`